Move VAOS unit tests: Batch 2
